+++ alembic/versions/9ae1382693cf_adds_knowledge_graphs.py
@@ -1,0 +1,68 @@
+"""adds knowledge graphs
+
+Revision ID: 9ae1382693cf
+Revises: e3e108cd4b22
+Create Date: 2025-12-29 08:24:46.628109
+
+"""
+
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import postgresql
+
+# revision identifiers, used by Alembic.
+revision = "9ae1382693cf"
+down_revision = "e3e108cd4b22"
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        "knowledge_graph",
+        sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
+        sa.Column("organization_id", postgresql.UUID(as_uuid=True), nullable=True),
+        sa.Column("project_id", postgresql.UUID(as_uuid=True), nullable=True),
+        sa.Column("created_by", postgresql.UUID(as_uuid=True), nullable=True),
+        sa.Column("created_at", sa.DateTime(), nullable=True),
+        sa.Column("name", sa.String(), nullable=True),
+        sa.Column("description", sa.String(), nullable=True),
+        sa.Column("type", sa.String(), nullable=True),
+        sa.ForeignKeyConstraint(["created_by"], ["user.id"], ondelete="SET NULL"),
+        sa.ForeignKeyConstraint(
+            ["organization_id"], ["organization.id"], ondelete="CASCADE"
+        ),
+        sa.ForeignKeyConstraint(["project_id"], ["project.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        op.f("ix_knowledge_graph_created_by"),
+        "knowledge_graph",
+        ["created_by"],
+        unique=False,
+    )
+    op.create_index(
+        op.f("ix_knowledge_graph_organization_id"),
+        "knowledge_graph",
+        ["organization_id"],
+        unique=False,
+    )
+    op.create_index(
+        op.f("ix_knowledge_graph_project_id"),
+        "knowledge_graph",
+        ["project_id"],
+        unique=False,
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f("ix_knowledge_graph_project_id"), table_name="knowledge_graph")
+    op.drop_index(
+        op.f("ix_knowledge_graph_organization_id"), table_name="knowledge_graph"
+    )
+    op.drop_index(op.f("ix_knowledge_graph_created_by"), table_name="knowledge_graph")
+    op.drop_table("knowledge_graph")
+    # ### end Alembic commands ###
